+++ core/src/main/java/be/orbinson/aem/dictionarytranslator/models/package-info.java
@@ -1,4 +1,4 @@
-@Version("1.0.0")
+@Version("2.0.0")
 package be.orbinson.aem.dictionarytranslator.models;
 
 import org.osgi.annotation.versioning.Version;