+++ core/src/test/java/be/orbinson/aem/dictionarytranslator/services/impl/DictionaryServiceImplTest.java
@@ -1,9 +1,11 @@
 package be.orbinson.aem.dictionarytranslator.services.impl;
 
+import be.orbinson.aem.dictionarytranslator.models.Dictionary;
 import be.orbinson.aem.dictionarytranslator.services.DictionaryService;
 import com.day.cq.replication.Replicator;
 import io.wcm.testing.mock.aem.junit5.AemContext;
 import io.wcm.testing.mock.aem.junit5.AemContextExtension;
+import org.apache.sling.api.resource.PersistenceException;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.junit.jupiter.api.BeforeEach;
@@ -12,11 +14,13 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
 
 @ExtendWith({AemContextExtension.class, MockitoExtension.class})
 class DictionaryServiceImplTest {
@@ -29,40 +33,65 @@ class DictionaryServiceImplTest {
     void setup() {
         context.registerService(Replicator.class, mock(Replicator.class));
         dictionaryService = context.registerInjectActivateService(new DictionaryServiceImpl());
+
+        context.load().json("/content.json", "/content");
     }
 
     @Test
-    void returnsUniqueDictionaries() {
-        context.create().resource("/content/dictionaries/site-a/i18/en", Map.of("jcr:language", "en"));
-        context.create().resource("/content/dictionaries/site-a/i18/fr", Map.of("jcr:language", "fr"));
-        context.create().resource("/content/dictionaries/site-b/i18/en", Map.of("jcr:language", "en"));
-
+    void dictionaryListShouldNotContainDuplicates() {
         ResourceResolver resourceResolver = spy(context.resourceResolver());
-        doReturn(
-                List.of(
-                        context.resourceResolver().getResource("/content/dictionaries/site-a/i18"),
-                        context.resourceResolver().getResource("/content/dictionaries/site-a/i18"),
-                        context.resourceResolver().getResource("/content/dictionaries/site-b/i18")
+
+        doReturn(List.of(
+                        context.resourceResolver().getResource("/content/dictionaries/fruit/i18n"),
+                        context.resourceResolver().getResource("/content/dictionaries/fruit/i18n"),
+                        context.resourceResolver().getResource("/content/dictionaries/vegetables/i18n")
                 ).iterator()
         ).when(resourceResolver).findResources(anyString(), anyString());
 
         List<Resource> dictionaries = dictionaryService.getDictionaries(resourceResolver);
 
         assertEquals(2, dictionaries.size());
-        assertEquals("/content/dictionaries/site-a/i18", dictionaries.get(0).getPath());
-        assertEquals("/content/dictionaries/site-b/i18", dictionaries.get(1).getPath());
+        assertEquals("/content/dictionaries/fruit/i18n", dictionaries.get(0).getPath());
+        assertEquals("/content/dictionaries/vegetables/i18n", dictionaries.get(1).getPath());
     }
 
     @Test
-    void returnsCorrectLanguages() {
-        context.create().resource("/content/dictionaries/site/i18");
-        context.create().resource("/content/dictionaries/site/i18/fr", Map.of("jcr:language", "fr"));
-        context.create().resource("/content/dictionaries/site/i18/rep:policy");
-        context.create().resource("/content/dictionaries/site/i18/en", Map.of("jcr:language", "en"));
+    void dictionaryServiceShouldReturnCorrectLanguages() {
+        Resource dictionary = context.currentResource("/content/dictionaries/fruit/i18n");
 
-        List<String> languages = dictionaryService.getLanguages(context.currentResource("/content/dictionaries/site/i18"));
+        assertEquals(List.of("en", "nl_BE"), dictionaryService.getLanguages(dictionary));
+    }
 
-        assertEquals("en", languages.get(0));
-        assertEquals("fr", languages.get(1));
+    @Test
+    void dictionaryServiceShouldBeAbleToAddNewLanguage() throws PersistenceException {
+        Resource dictionaryResource = context.currentResource("/content/dictionaries/fruit/i18n");
+
+        dictionaryService.addLanguage(dictionaryResource, "fr", "/content/dictionaries/fruit/i18n");
+
+        assertEquals(List.of("en", "fr", "nl_BE"), dictionaryService.getLanguages(dictionaryResource));
+    }
+
+    @Test
+    void dictionaryServiceShouldAddCorrectBasenameIfEmptry() throws PersistenceException {
+        Resource dictionaryResource = context.currentResource("/content/dictionaries/fruit/i18n");
+
+        dictionaryService.addLanguage(dictionaryResource, "fr", "");
+
+        Dictionary dictionary = context.request().adaptTo(Dictionary.class);
+
+        assertEquals("/content/dictionaries/fruit/i18n", dictionary.getBasename());
+    }
+
+    @Test
+    void dictionaryServiceShouldBeAbleToCreateDictionary() throws PersistenceException {
+        Resource parent = context.currentResource("/content/dictionaries");
+
+        dictionaryService.createDictionary(parent, "meat", new String[]{"en", "ar"}, null);
+
+        context.currentResource("/content/dictionaries/meat/i18n");
+        Dictionary dictionary = context.request().adaptTo(Dictionary.class);
+
+        assertNotNull(dictionary);
+        assertEquals(List.of("ar", "en"), dictionary.getLanguages());
     }
 }