+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Only single dictionary can be selected [#62](https://github.com/orbinson/aem-dictionary-translator/issues/62)
+- Last modified date is remove as it was not correct
 
 ## [1.2.1] - 2024-11-26